cnn-service/requirements.txt
email-validator==1.1.2
fastapi==0.81.0
PyJWT==1.7.1
python-decouple==3.3
uvicorn==0.12.3
tensorflow-macos
aiokafka==0.7.2
pillow==9.2.0
numpy==1.23.2
python-multipart==0.0.5
